+++ coco/src/components/Header.jsx
@@ -4,6 +4,12 @@ import { useMediaQuery } from 'react-responsive';
 import { useNavigate } from 'react-router-dom';
 
 export const Header = () => {
+  const navigate = useNavigate();
+
+  const goToLogin = () => {
+    navigate('/login');
+  }
+  
   return (
     <div className="navbar">
       <h2>COCO : Coding Coach</h2>
@@ -13,7 +19,7 @@ export const Header = () => {
         <h3>게시판</h3>
       </div>
 
-      <div className="login">
+      <div className="login" onClick={goToLogin}>
         <h3>LOGIN</h3>
       </div>
     </div>
