+++ coco/src/components/SignUp.css
@@ -1,93 +1,100 @@
-/* 회원가입 */
-.signUpForm{
-    display: grid;
-    grid-template-columns: 2fr 1fr;
+.loginContainer{
+    width: 500px;
 }
 
-
-.signUpRight{
-    border: 4px outset gray;
-    padding: 20px;
-    border-radius: 10px;
-}
-
-.signUpRight > h3{
-    text-align: center;
-    padding-bottom: 20px;
-}
-
-/* 각각 박스(아이디, 이름, ...) */
-.signUpBox{
-    padding: 10px;
+.loginForm{
+    width: 100%;
     display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    margin: 20px 0px;
+    flex-direction: column;
+    flex-wrap: nowrap;
+    justify-content: space-evenly;
+    align-content: center;
+    border-bottom: 1px solid rgb(224, 224, 224);
+    border-left: 1px solid  rgb(224, 224, 224);
+    border-right: 1px solid  rgb(224, 224, 224);
 }
 
-.signUpBox > input{
-    width: 90%;
-    border:none;
-    background-color: rgb(214, 214, 214);
+.loginBox{
+    margin: 20px 10px;
+    /* padding: 10px; */
+    display: flex;
+    flex-direction: column;
+    flex-wrap: nowrap;
+    /* align-items: center; */
+}
+
+.loginBox :nth-child(1){
+    margin-bottom: 10px;
+}
+
+.loginBox>input{
+    width: 100%;
+    border: none;
     padding: 10px 5px;
-    font-size: medium;
+    background-color: rgb(227, 227, 227);
     border-radius: 5px;
 }
 
-/* 아이디 입력 */
-.signUpID{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-.signUpID > input{
-    width: 70%;
-    border:none;
-    background-color: rgb(214, 214, 214);
-    padding-right: 10px;
-}
-
-/* 학생or선생 */
-.signUpID > button{
-    font-size: 1vw;
+.loginBox>input::after{
     border: none;
-    color: rgb(141, 140, 140);
 }
 
-.signUpTypes{
-    display: flex;
-    flex-flow: row;
-    justify-content: space-evenly;
-    margin: 20px 0px;
+.loginConfirm > Button{
+    width: 80%;
+    background-color: gray;
+    color: white;
+    /* align-content: center; */
 }
 
-.signUpTypes > div > label{
-    padding-right: 10px;
-}
-
-/* 가입버튼 */
-.signUpConfirm{
-    padding: 10px;
+.loginConfirm{
     display: flex;
     flex-direction: row;
     justify-content: center;
 }
-.signUpConfirmBtn{
-    background-color: gray;
-    color: white;
-    width: 80%;
+
+.loginOptions{
+    font-size: small;
+    text-decoration: underline;
+    color: rgb(109, 150, 167);
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
 }
 
-/* 로그인 */
-.loginForm{
-    width: 500px;
-    border: 4px outset gray;
-    padding: 20px;
-    border-radius: 10px;
+.loginOptions > span, .loginFind:hover{
+    cursor: pointer;
 }
 
-.loginForm > h3{
-    text-align: center;
-    padding-bottom: 20px;
+.checkID{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.checkID > input{
+    width: 100%;
+    border: none;
+    padding: 10px 5px;
+    background-color: rgb(227, 227, 227);
+    border-radius: 5px;
+    margin-bottom: 0 !important;
+}
+
+.checkID > button{
+    font-size: small;
+    margin-left: 20px;
+    width: 100px;
+    border: 1px solid gray;
+    border-radius: 5px;
+}
+
+.signUpTypes{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+}
+
+.signUpType{
+    margin: 0 !important;
 }
